fix: update variable parsing by looking at number of opening/closing parentheses
Within a Snakemake rule, the input and output sections can include functions with key-named variables, introducing equal signs within their values. In the past, our implementation relied on equal signs as indicators to identify the end of a variable. However, this approach became problematic when dealing with key-named variables, as it led to invalid parsing.

With the introduction of the new approach, we've addressed this challenge. Instead of relying on equal signs, the new method utilizes a count of opening and closing parentheses. This count is used to accurately determine when all information for a variable, including those with key names, has been successfully read.